requirements.txt
fastapi
uvicorn[standard]
ytmusicapi
spotipy
supabase
openai
python-dotenv
azure-functions
